services/bounty-agent/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
jinja2>=3.1.3
pytest>=8.0.0
pytest-asyncio>=0.23.0